Time: 12 ms (70.48%), Space: 17.3 MB (52.35%) - LeetHub

+++ 662-maximum-width-of-binary-tree/662-maximum-width-of-binary-tree.cpp
@@ -34,8 +34,8 @@ public:
                 if(i == 0) first = cur_id ; 
                 if(i == size-1) last = cur_id ;
                 
-                if(tmp->left) nodes.push({tmp->left, 2LL*cur_id+1}) ; //LL-->working like long long int
-                if(tmp->right) nodes.push({tmp->right, 2LL*cur_id+2}) ;
+                if(tmp->left) nodes.push({tmp->left, (long long)2*cur_id+1}) ; //LL-->working like long long int
+                if(tmp->right) nodes.push({tmp->right, (long long)2*cur_id+2}) ;
             }
             ans = max(ans, last-first+1) ;
         }
